requirements.txt
fastapi
uvicorn
tensorflow
scikit-learn
numpy
geopy
requests
python-dotenv